backend/postapi/requirements.txt
Django==3.1.7
djangorestframework==3.12.2
dj-rest-auth==1.0.4
django-allauth==0.44.0
django-cors-headers==3.7.0